requirements.txt.txt
paho-mqtt>=1.6.1
numpy
pandas
scikit-learn
matplotlib
joblib